# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoTriage Intake Service",
    "description": "REST API for ticket submission, storage, and retrieval.",
    "version": "1.0.0"
  },
  "paths": {
    "/tickets": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Submit a new support ticket",
        "description": "Accept a new ticket and persist it to the database.\n\nReturns the generated `ticket_id` and an acknowledgment flag.",
        "operationId": "create_ticket_tickets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "List tickets with optional filters",
        "description": "Return a paginated list of tickets.\n\nSupports optional filtering by `source`, `status`, and a `date_from`/`date_to`\nrange applied to the ticket's `timestamp` field.",
        "operationId": "list_tickets_tickets_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source system",
              "title": "Source"
            },
            "description": "Filter by source system"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include tickets at or after this ISO-8601 datetime",
              "title": "Date From"
            },
            "description": "Include tickets at or after this ISO-8601 datetime"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include tickets at or before this ISO-8601 datetime",
              "title": "Date To"
            },
            "description": "Include tickets at or before this ISO-8601 datetime"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (open, in_progress, resolved, closed)",
              "title": "Status"
            },
            "description": "Filter by status (open, in_progress, resolved, closed)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max records to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max records to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of records to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Retrieve a single ticket by ID",
        "description": "Fetch a single ticket by its UUID.\n\nReturns 404 if the ticket does not exist.",
        "operationId": "get_ticket_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginationMeta"
      },
      "TicketCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "description": "Short title of the ticket"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "description": "Full description of the issue"
          },
          "source": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source",
            "description": "Origin system or team (e.g. 'web', 'mobile', 'ops')"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "ISO-8601 datetime when the ticket was originally created"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "source",
          "timestamp"
        ],
        "title": "TicketCreate"
      },
      "TicketCreateResponse": {
        "properties": {
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Ticket received and stored successfully"
          }
        },
        "type": "object",
        "required": [
          "ticket_id"
        ],
        "title": "TicketCreateResponse"
      },
      "TicketResponse": {
        "properties": {
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "resolved",
              "closed"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "title",
          "description",
          "source",
          "timestamp",
          "status",
          "created_at"
        ],
        "title": "TicketResponse"
      },
      "TicketsListResponse": {
        "properties": {
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/TicketResponse"
            },
            "type": "array",
            "title": "Tickets"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "tickets",
          "pagination"
        ],
        "title": "TicketsListResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}